fixed is_newest and added patch checks
The function is_newest would try to check for an json that DNE. So, when initializing the Version class, check_allowed is called to set self.allowed: bool which will be checked when certain functions run. This stops the is_newest crash by exiting the function before it calls a non-existing json. If 'patch' is found in name of a release, a message displays that a patch is available. Print statements are left in for when testing is created.