Window resize, schedule editing, PEP 8 Style
This is an early change from 0.3.8. Skip feature still not ready, but these bug fixes are.

Resizing the session window crashes when the next image is loaded. The QSize.scaled() parameters were not correctly formatted.

After closing a session, removing rows, then starting a new session will result in an error. self.schedule was not updated accordingly.

Revised code to adhere to PEP8.