Scheduling bugs, refactor, and preset autocomplete
Schedule editing functions would not work properly. Functions involving the schedule are now revised with an unenabled QTableWidgetItemFlag and more limited conditions.
Some code involved with scheduling refactored in the process.

New signal is connected to self.preset_loader_box that aids in debugging, and adds an autocomplete feature for presets.