Adds initialisation step for selection data
Issue: When no recent config is available, the folders and files keys in the self.selection of MainApp are not defined. This results in a crash when attempting to add folders or files.
Fix: Created a data initialisation stage for app startup. Ensures the data is initialised, then tries to load config data if it exists.